requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.11.9
python-multipart==0.0.18
grpcio==1.75.1
grpcio-tools==1.75.1
requests==2.32.5
python-dotenv==1.1.1
aiohttp==3.12.15